refactor: :recycle: removed unecessary classes folder that led to weird nesting
